feat: adding Featured Speaker task
Adding task for featured speakers, so that it memcache gets updated when
a new session is added. If the new session is done by a speaker that has
more than one session at that Conference, they become the featured
speaker for all of Conference Central.

Also, adding code for how to solve the inequality question.